Rename AirSpecRunner -> AirSpecRunnerSbt
